Added Scheduler Queues and Time
This commit was honestly a little too big. It comes with various queues for inter-task communication and an interface for getting and incrementing the internal time.

+++ hydra-core/src/task.rs
@@ -15,7 +15,7 @@ pub struct Task {
     pub name: &'static str,
     pub state: State,
     priority: usize,
-    pub last_ran: usize,
+    pub last_ran: u64, 
     predicate: fn(&Task, &QueueMap) -> bool,
     func: fn(&mut Task, &mut QueueMap) -> (),
 }
